:bug: Fix transaction_id parsing.
Transaction IDs logged as part of the resilient client User-Agent HTTP
header tend to abut a parenthesis and a quote mark directly, but these
non-whitespace characters should not be extracted into the
transaction_id value that is logged. I've amended the pattern to
accommodate uppercase, lowercase, numbers, hyphen, underscore and colon
and rationalised the test cases. As far as I'm aware these are the only
characters that should appear in a transaction id (the colon is a quirk
of the logging in PAM).