Merge from master (#39)
* Removing restrictions on PAM tids

* Improving regex for transaction IDs from unstructured logs (i.e. notifications-rw).

* Supporting extracting key-value pairs from a json message

* Removing /health and /gtg entries in order to temporary reduce the logging volume

* make json logs accept any value

* :bug: Fix transaction_id parsing.

Transaction IDs logged as part of the resilient client User-Agent HTTP
header tend to abut a parenthesis and a quote mark directly, but these
non-whitespace characters should not be extracted into the
transaction_id value that is logged. I've amended the pattern to
accommodate uppercase, lowercase, numbers, hyphen, underscore and colon
and rationalised the test cases. As far as I'm aware these are the only
characters that should appear in a transaction id (the colon is a quirk
of the logging in PAM).

* Remove mdc related values from logs

* simple code for unwanted fields

* Add dummy message

* Fix Dockerfile

* remove dummy message

* add cluster status field for monitoring events

* test cleanup

* cleanup

* Updated circle ci

* New line

* fixed circle ci build